ig8g
Changed class T to int

+++ SuperSorter/SuperSorter/main.cpp
@@ -8,7 +8,7 @@ void printArray(int arr[], int n);
 void heapSort(int arr[], int n);			// main function to do heap sort
 void heapify(int arr[], int n, int i);
 
-void mergesort(int a[], size_t l, size_t h) {
+void mergesort(int a[], int l, int h) {
 	if (h - l == 1) {
 		if (a[l] > a[h]) {
 			int t = a[l];
@@ -20,14 +20,14 @@ void mergesort(int a[], size_t l, size_t h) {
 
 	}
 	else if (h > l) {
-		size_t size = h - l + 1;
-		size_t m = l + (h - l) / 2;
+		int size = h - l + 1;
+		int m = l + (h - l) / 2;
 		mergesort(a, l, m);
 		mergesort(a, m + 1, h);
 		int* b = new int[size];
-		size_t k = 0;
-		size_t i = l;
-		size_t j = m + 1;
+		int k = 0;
+		int i = l;
+		int j = m + 1;
 		while (i <= m && j <= h)
 			if (a[i] <= a[j])
 				b[k++] = a[i++];
@@ -47,7 +47,7 @@ void mergesort(int a[], size_t l, size_t h) {
 	}
 }
 
-void mergesort(int a[], size_t size) {
+void mergesort(int a[], int size) {
 	mergesort(a, 0, size - 1);
 }
